Review By Gaoxiang Zhang, classmate in CS5610
I must say, your website is very creative and functional. This system of filtering meals inspired me, because my last project was also an ordering website. Glad to see that you fixed this filter function. I have a small suggestion for the meal page, is that in my opinion, it is better to use css make all unfilter list of food in a div with scroll display, so that it will not make your page so long in the vertical. You can do this during iteration version. 
Say that again, you have done really good job, thanks to bring that to me. Hope to see your next version.